TransformSanitise: Merge tests for sanitise transformation
